refactor(admin): move LLM factory + test runner to infrastructure behind ports (#19)
