+++ backend/src/hiresense/admin/domain/__init__.py
@@ -7,16 +7,15 @@ from hiresense.admin.domain.feature_registry import (
     get_feature,
 )
 from hiresense.admin.domain.llm_config_service import LLMConfigService
-from hiresense.admin.domain.llm_factory import LLMFactory, UnsupportedProviderError
 from hiresense.admin.domain.llm_settings_service import (
     GlobalSettingsView,
     LLMSettingsService,
     LLMSettingsServiceError,
 )
-from hiresense.admin.domain.llm_test_runner import LLMTestRunner, TestResult
 from hiresense.admin.domain.masking import mask_api_key
 from hiresense.admin.domain.pricing import DEFAULT_PRICING, ModelPricing, estimate_cost_usd
 from hiresense.admin.domain.resolved_config import ResolvedConfig
+from hiresense.admin.domain.test_result import TestResult
 from hiresense.admin.domain.usage_aggregator import DashboardSummary, UsageAggregator
 from hiresense.admin.domain.usage_recorder import UsageRecorder
 
@@ -30,14 +29,11 @@ __all__ = [
     "FeatureDescriptor",
     "GlobalSettingsView",
     "LLMConfigService",
-    "LLMFactory",
     "LLMSettingsService",
     "LLMSettingsServiceError",
-    "LLMTestRunner",
     "ModelPricing",
     "ResolvedConfig",
     "TestResult",
-    "UnsupportedProviderError",
     "UsageAggregator",
     "UsageRecorder",
     "all_feature_keys",
