+++ backend/src/hiresense/admin/infrastructure/__init__.py
@@ -3,12 +3,14 @@ from hiresense.admin.infrastructure.feature_configured_llm_adapter import (
 )
 from hiresense.admin.infrastructure.llm_audit_log_model import LLMAuditLog
 from hiresense.admin.infrastructure.llm_audit_log_repository import LLMAuditLogRepository
+from hiresense.admin.infrastructure.llm_factory import LLMFactory, UnsupportedProviderError
 from hiresense.admin.infrastructure.llm_feature_override_model import LLMFeatureOverride
 from hiresense.admin.infrastructure.llm_feature_override_repository import (
     LLMFeatureOverrideRepository,
 )
 from hiresense.admin.infrastructure.llm_settings_model import LLMSettings
 from hiresense.admin.infrastructure.llm_settings_repository import LLMSettingsRepository
+from hiresense.admin.infrastructure.llm_test_runner import LLMTestRunner
 from hiresense.admin.infrastructure.llm_usage_log_model import LLMUsageLog
 from hiresense.admin.infrastructure.llm_usage_log_repository import (
     LLMUsageLogRepository,
@@ -23,12 +25,15 @@ __all__ = [
     "FeatureConfiguredLLMAdapter",
     "LLMAuditLog",
     "LLMAuditLogRepository",
+    "LLMFactory",
     "LLMFeatureOverride",
     "LLMFeatureOverrideRepository",
     "LLMSettings",
     "LLMSettingsRepository",
+    "LLMTestRunner",
     "LLMUsageLog",
     "LLMUsageLogRepository",
+    "UnsupportedProviderError",
     "UsageBucket",
     "UsageTotals",
     "UsageTrackingLLMAdapter",
