+++ backend/src/hiresense/admin/ports/__init__.py
@@ -1,14 +1,18 @@
-"""Admin module ports — repository protocols for the LLM-config services."""
+"""Admin module ports — repository protocols and service ports for the LLM-config services."""
 from hiresense.admin.ports.llm_audit_log_repository_port import LLMAuditLogRepositoryPort
+from hiresense.admin.ports.llm_factory_port import LLMFactoryPort
 from hiresense.admin.ports.llm_feature_override_repository_port import (
     LLMFeatureOverrideRepositoryPort,
 )
 from hiresense.admin.ports.llm_settings_repository_port import LLMSettingsRepositoryPort
+from hiresense.admin.ports.llm_test_runner_port import LLMTestRunnerPort
 from hiresense.admin.ports.llm_usage_log_repository_port import LLMUsageLogRepositoryPort
 
 __all__ = [
     "LLMAuditLogRepositoryPort",
+    "LLMFactoryPort",
     "LLMFeatureOverrideRepositoryPort",
     "LLMSettingsRepositoryPort",
+    "LLMTestRunnerPort",
     "LLMUsageLogRepositoryPort",
 ]
