+++ backend/src/hiresense/bootstrap/admin.py
@@ -7,16 +7,16 @@ from hiresense.admin.api import AdminProvider
 from hiresense.admin.domain import (
     APIKeyCipher,
     LLMConfigService,
-    LLMFactory,
     LLMSettingsService,
-    LLMTestRunner,
     UsageAggregator,
     UsageRecorder,
 )
 from hiresense.admin.infrastructure import (
     LLMAuditLogRepository,
+    LLMFactory,
     LLMFeatureOverrideRepository,
     LLMSettingsRepository,
+    LLMTestRunner,
     LLMUsageLogRepository,
 )
 from hiresense.bootstrap.shared_infra import SharedInfra
